[Discord] Add UseDiscordNameField setting
If set to "True", the discord "username" field will be used with relayed
messages (this is the old behavior of the script). It looks better in
the discord client but can result in some issues with message ordering.
When not enabled, usernames are surrounded with angled brackets <> and
put before the message (similar to the bot window).